pyproject.toml
[project]
name = "intelligent-document-sorter"
version = "0.1.0"
description = "An agent that sorts documents intelligently using Composio + OpenAI."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "composio>=0.8.15",
    "composio-gemini>=0.8.13",
    "composio-google>=0.8.13",
    "dotenv>=0.9.9",
    "google-genai>=1.38.0",
    "google-generativeai>=0.8.5",
    "typer>=0.19.1",
]
